refactor: move and refactor vibrations to feature dir
