--- lib/sound/mixins/vibrations.dart
@@ -1,26 +1,0 @@
-import 'dart:async';
-
-import 'package:vibration/vibration.dart';
-
-class Vibrations {
-  late final bool _customVibrationsAvailable;
-
-  Future<bool> init() async {
-    try {
-      final bool hasBasicVibrator = await Vibration.hasVibrator() ?? false;
-      final bool hasExtenderVibrator = await Vibration.hasCustomVibrationsSupport() ?? false;
-
-      return _customVibrationsAvailable = hasBasicVibrator && hasExtenderVibrator;
-    } catch (_) {
-      return _customVibrationsAvailable = false;
-    }
-  }
-
-  bool vibrate({int durationInMs = 150}) {
-    if (_customVibrationsAvailable) {
-      unawaited(Vibration.vibrate(duration: durationInMs));
-    }
-
-    return _customVibrationsAvailable;
-  }
-}
